Embed PSQL and MYSQL clients in plugin & Use python buildpack
The previous iteration of this plugin used a custom buildpack.
However, it became brittle and it made the program to actually have to
install the aws (and not the buildpack). Instead, we will use the
native python buildpack and use pip to install any requirements that
were broken in the last iteration of the plugin. The one requirement
that failed to install manually was awscli. Instead, we now have a
requirements.txt to specify the need for it.

In addition, the previous buildpack had the mysql and psql tools.
Instead of using that buildpack, we will embed the tools in the plugin
itself and will deploy with the export/import app bits