Adds support for flags in the upload-backup-data command
This changeset adds two flags to the `upload-backup-data` command:

* `-p` - Used to specify the file path to upload.
* `-store` - Used to specify the name of the S3 service to send the upload to.

Note that using the `-store` flag will cause the `upload-backup-data` command to run in a non-interactive mode and will not prompt the user to choose an S3 bucket to target.

h/t to @jcscottiii for all of the help and mentoring with Go!